requirements.txt
Pillow
numpy
PyYAML
matplotlib
cartopy
